Adds a groupcache-base cache, sort of.
GroupCache has the concept of a backfill, whereby it will pull
from another source on cache miss, helping reduce hotspots and
distributing the pull load. As-is, this isn't compatible with
the []Cache concept, so the backfiller is nil, which will work,
but results in worse than usual performance as a cache miss
will result in all nodes being checked instead of the correct
member pulling it through.