Adds a groupcache-based cache with caveats, PoC.
This reimplementation will create a cache from the default config
if one doesn't exist when a tile is set. This results in a locked
map check on every set. I don't like it, but it's a PoC. Fillfuncs
are nil, so one of the larger value-props for groupcache is moot
until that is implmented.